moved equaliser GUI functions to eq.cpp
