+++ src/main.cpp
@@ -3679,149 +3679,6 @@ void MainFrame::startRxStream()
 }
 
 
-#define SBQ_MAX_ARGS 5
-
-void *MainFrame::designAnEQFilter(const char filterType[], float freqHz, float gaindB, float Q)
-{
-    char  *arg[SBQ_MAX_ARGS];
-    char   argstorage[SBQ_MAX_ARGS][80];
-    void  *sbq;
-    int    i, argc;
-
-    assert((strcmp(filterType, "bass") == 0)   ||
-           (strcmp(filterType, "treble") == 0) ||
-           (strcmp(filterType, "equalizer") == 0));
-
-    for(i=0; i<SBQ_MAX_ARGS; i++) {
-        arg[i] = &argstorage[i][0];
-    }
-
-    argc = 0;
-
-    if ((strcmp(filterType, "bass") == 0) || (strcmp(filterType, "treble") == 0)) {
-        sprintf(arg[argc++], "%s", filterType);
-        sprintf(arg[argc++], "%f", gaindB+1E-6);
-        sprintf(arg[argc++], "%f", freqHz);
-    }
-
-    if (strcmp(filterType, "equalizer") == 0) {
-        sprintf(arg[argc++], "%s", filterType);
-        sprintf(arg[argc++], "%f", freqHz);
-        sprintf(arg[argc++], "%f", Q);
-        sprintf(arg[argc++], "%f", gaindB+1E-6);
-    }
-
-    assert(argc <= SBQ_MAX_ARGS);
-    // Note - the argc count doesn't include the command!
-    sbq = sox_biquad_create(argc-1, (const char **)arg);
-    assert(sbq != NULL);
-
-    return sbq;
-}
-
-void  MainFrame::designEQFilters(paCallBackData *cb)
-{
-    // init Mic In Equaliser Filters
-
-    if (m_newMicInFilter) {
-        //printf("designing new Min In filters\n");
-        cb->sbqMicInBass   = designAnEQFilter("bass", wxGetApp().m_MicInBassFreqHz, wxGetApp().m_MicInBassGaindB);
-        cb->sbqMicInTreble = designAnEQFilter("treble", wxGetApp().m_MicInTrebleFreqHz, wxGetApp().m_MicInTrebleGaindB);
-        cb->sbqMicInMid    = designAnEQFilter("equalizer", wxGetApp().m_MicInMidFreqHz, wxGetApp().m_MicInMidGaindB, wxGetApp().m_MicInMidQ);
-    }
-
-    // init Spk Out Equaliser Filters
-
-    if (m_newSpkOutFilter) {
-        //printf("designing new Spk Out filters\n");
-        //printf("designEQFilters: wxGetApp().m_SpkOutBassFreqHz: %f\n",wxGetApp().m_SpkOutBassFreqHz);
-        cb->sbqSpkOutBass   = designAnEQFilter("bass", wxGetApp().m_SpkOutBassFreqHz, wxGetApp().m_SpkOutBassGaindB);
-        cb->sbqSpkOutTreble = designAnEQFilter("treble", wxGetApp().m_SpkOutTrebleFreqHz, wxGetApp().m_SpkOutTrebleGaindB);
-        cb->sbqSpkOutMid    = designAnEQFilter("equalizer", wxGetApp().m_SpkOutMidFreqHz, wxGetApp().m_SpkOutMidGaindB, wxGetApp().m_SpkOutMidQ);
-    }
-}
-
-void  MainFrame::deleteEQFilters(paCallBackData *cb)
-{
-    if (m_newMicInFilter) {
-        sox_biquad_destroy(cb->sbqMicInBass);
-        sox_biquad_destroy(cb->sbqMicInTreble);
-        sox_biquad_destroy(cb->sbqMicInMid);
-    }
-    if (m_newSpkOutFilter) {
-        sox_biquad_destroy(cb->sbqSpkOutBass);
-        sox_biquad_destroy(cb->sbqSpkOutTreble);
-        sox_biquad_destroy(cb->sbqSpkOutMid);
-    }
-}
-
-// returns number of output samples generated by resampling
-int resample(SRC_STATE *src,
-            short      output_short[],
-            short      input_short[],
-            int        output_sample_rate,
-            int        input_sample_rate,
-            int        length_output_short, // maximum output array length in samples
-            int        length_input_short
-            )
-{
-    SRC_DATA src_data;
-    float    input[length_input_short];
-    float    output[length_output_short];
-    int      ret;
-
-    assert(src != NULL);
-
-    src_short_to_float_array(input_short, input, length_input_short);
-
-    src_data.data_in = input;
-    src_data.data_out = output;
-    src_data.input_frames = length_input_short;
-    src_data.output_frames = length_output_short;
-    src_data.end_of_input = 0;
-    src_data.src_ratio = (float)output_sample_rate/input_sample_rate;
-
-    ret = src_process(src, &src_data);
-    assert(ret == 0);
-
-    assert(src_data.output_frames_gen <= length_output_short);
-    src_float_to_short_array(output, output_short, src_data.output_frames_gen);
-
-    return src_data.output_frames_gen;
-}
-
-
-// Decimates samples using an algorithm that produces nice plots of
-// speech signals at a low sample rate.  We want a low sample rate so
-// we don't hammer the graphics system too hard.  Saves decimated data
-// to a fifo for plotting on screen.
-
-void resample_for_plot(struct FIFO *plotFifo, short buf[], int length, int fs)
-{
-    int decimation = fs/WAVEFORM_PLOT_FS;
-    int nSamples, sample;
-    int i, st, en, max, min;
-    short dec_samples[length];
-
-    nSamples = length/decimation;
-
-    for(sample = 0; sample < nSamples; sample += 2)
-    {
-        st = decimation*sample;
-        en = decimation*(sample+2);
-        max = min = 0;
-        for(i=st; i<en; i++ )
-        {
-            if (max < buf[i]) max = buf[i];
-            if (min > buf[i]) min = buf[i];
-        }
-        dec_samples[sample] = max;
-        dec_samples[sample+1] = min;
-    }
-    codec2_fifo_write(plotFifo, dec_samples, nSamples);
-}
-
-
 //---------------------------------------------------------------------------------------------
 // Main real time procesing for tx and rx of FreeDV signals, run in its own thread
 //---------------------------------------------------------------------------------------------
